requirements.txt
streamlit
pytesseract
spacy
google-cloud-storage
Pillow
pdfminer.six
pdf2image
